Change way of skipping the first 4 primes
The conjecture of Firoobakht does not apply for the first 4 prime
numbers. For this reason, when calculating the next prime number to a
given number (let's say 2), before applying the conjecture to find the
gap between the next number, I need to see if it is one of these first
numbers. If it is, then, I have a table to find the next.