Change way of finding the next prime.
The conjecture of Firoozbakht (given a prime) finds a number that
demonstrates that is bigger (but closer enough) to the gap betwen the
given prime and its subsequent.
If used, this number as a upper boundary, it could result (as I was
using it), that some prime numbers may be skipped for time to time.
For instance applying the formula, we know that:
   log(pn)^2 - log(pn) > gn
In the case of 97, this would be:
   floor( log(97)^2 - log(97) ) = 16
Certainly 113 is a prime, and also 109, 107, 103 and 101. So using the
algorithm as I was using it, I jumped in some cases quite a bunch of
primes, which is not valid for the competition (the primes should be
subsequent).