Structural changes in the algorithm
Now I have a single dispatcher that decides what needs to be evaluated
as a prime, and several workers which evaluates if the number is a
prime number or not. If it is a prime number, then, the program will
enqueue the number in the sorting job, which will sort numbers every
once in a while.
At last, when the numbers are sorted, they will be handled to a printer
job.